turn files from PascalCase into kebab-case
