+++ packages/functions/src/get-context.ts
@@ -1,0 +1,13 @@
+type Context = {
+  waitUntil?: (promise: Promise<unknown>) => void;
+  headers?: Record<string, string>;
+};
+
+export const SYMBOL_FOR_REQ_CONTEXT = Symbol.for('@vercel/request-context');
+
+export function getContext(): Context {
+  const fromSymbol: typeof globalThis & {
+    [SYMBOL_FOR_REQ_CONTEXT]?: { get?: () => Context };
+  } = globalThis;
+  return fromSymbol[SYMBOL_FOR_REQ_CONTEXT]?.get?.() ?? {};
+}
