+++ packages/functions/src/index.ts
@@ -1,3 +1,4 @@
 export { geolocation, ipAddress } from './headers';
 export { getEnv } from './get-env';
 export { waitUntil } from './wait-until';
+export { getVercelOidcToken, awsCredentialsProvider } from './oidc';
