+++ packages/functions/src/oidc/index.ts
@@ -1,2 +1,2 @@
-export * from './awsCredentialsProvider';
-export * from './getVercelOidcToken';
+export * from './aws-credentials-provider';
+export * from './get-vercel-oidc-token';
